requirements.txt
# Core FastAPI Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Caching & Session Management
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI & ML Services
openai==1.3.7
anthropic==0.7.8
google-cloud-vision==3.4.5
pinecone-client==2.2.4
langchain==0.0.350
langgraph==0.0.20

# HTTP Client & External APIs
httpx==0.25.2
aiohttp==3.9.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# WebSocket Support
websockets==12.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

# Environment & Configuration
python-dotenv==1.0.0
